move telemetry helper for validator
